Separate the input data and output path; unzip json in Python
