+++ .gitignore
@@ -1,1 +1,3 @@
 .idea/
+output/
+data/__MACOSX
