--- data/Readme.txt
@@ -1,8 +1,0 @@
-11/13
-
-Reorganized the code. I will update my code based this version later.
-
-
-At present， 2000+ substations cannot be found in graph. I'm checking why it happened.
-
-
